expand message type
